requirements.txt
pandas
numpy
yfinance
requests
pyyaml
apscheduler
beautifulsoup4
pyarrow
jinja2
pandas_market_calendars
